+++ dsmr_datalogger/admin.py
@@ -18,8 +18,26 @@ from .models.statistics import MeterStatistics
 class DataloggerSettingsAdmin(SingletonModelAdmin):
     fieldsets = (
         (
-            _('Protocol'), {
-                'fields': ['dsmr_version'],
+            _('Input method and DSMR protocol'), {
+                'fields': ['input_method', 'dsmr_version'],
+                'description': _(
+                    'Note: You might have to restart the "dsmr_datalogger" process for any changes to apply. '
+                    'See the <a href="https://dsmr-reader.readthedocs.io/nl/v4/faq/restart_processes.html">FAQ</a>.'
+                )
+            }
+        ),
+        (
+            _('When using serial socket input method'), {
+                'fields': ['serial_port'],
+                'description': _(
+                    'Note: You might have to restart the "dsmr_datalogger" process for any changes to apply. '
+                    'See the <a href="https://dsmr-reader.readthedocs.io/nl/v4/faq/restart_processes.html">FAQ</a>.'
+                )
+            }
+        ),
+        (
+            _('When using network socket input method'), {
+                'fields': ['network_socket_address', 'network_socket_port'],
                 'description': _(
                     'Note: You might have to restart the "dsmr_datalogger" process for any changes to apply. '
                     'See the <a href="https://dsmr-reader.readthedocs.io/nl/v4/faq/restart_processes.html">FAQ</a>.'
@@ -28,16 +46,7 @@ class DataloggerSettingsAdmin(SingletonModelAdmin):
         ),
         (
             _('Advanced'), {
-                'fields': ['com_port', 'process_sleep'],
-                'description': _(
-                    'Note: You will have to restart the "dsmr_datalogger" process for any changes to apply. '
-                    'See the <a href="https://dsmr-reader.readthedocs.io/nl/v4/faq/restart_processes.html">FAQ</a>.'
-                )
-            }
-        ),
-        (
-            _('Debugging'), {
-                'fields': ['log_telegrams'],
+                'fields': ['process_sleep', 'log_telegrams'],
             }
         ),
     )