+++ dsmr_datalogger/management/commands/dsmr_datalogger.py
@@ -5,6 +5,7 @@ from django.utils.translation import gettext as _
 
 from dsmr_backend.mixins import InfiniteManagementCommandMixin
 import dsmr_datalogger.services.datalogger
+from dsmr_datalogger.exceptions import InvalidTelegramError
 from dsmr_datalogger.models.settings import DataloggerSettings
 
 
@@ -15,8 +16,17 @@ class Command(InfiniteManagementCommandMixin, BaseCommand):
     help = _('Performs an DSMR P1 telegram reading on the serial port.')
     name = __name__  # Required for PID file.
 
+    # Persistent 'connection' to either the serial port or network socket, wrapped as generator.
+    telegram_generator = None
+
     def initialize(self):
         self.sleep_time = DataloggerSettings.get_solo().process_sleep
+        self.telegram_generator = dsmr_datalogger.services.datalogger.get_telegram_generator()
 
     def run(self, **options):
-        dsmr_datalogger.services.datalogger.read_and_process_telegram()
+        telegram = next(self.telegram_generator)
+
+        try:
+            dsmr_datalogger.services.datalogger.telegram_to_reading(data=telegram)
+        except InvalidTelegramError:
+            pass