+++ dsmr_datalogger/scripts/dsmr_datalogger_api_client.py
@@ -2,44 +2,25 @@
     https://dsmr-reader.readthedocs.io/en/latest/installation/datalogger.html
 
     Installation:
-        pip3 install pyserial==3.4 requests==2.22.0
+        pip3 install pyserial==3.4 requests==2.24.0 python-decouple==3.3
 """
 import datetime
 import logging
+import select
+import socket
 import time
+import re
 
 import serial
 import requests
+import decouple
 
 
-"""
-    These settings are only used when using this script as a dedicated remote datalogger.
-"""
-SLEEP = 0.5
-SERIAL_PORT = '/dev/ttyUSB0'
-SERIAL_BAUDRATE = 115200
-
-SERIAL_SETTINGS = dict(
-    port=SERIAL_PORT,
-    baudrate=SERIAL_BAUDRATE,
-    bytesize=serial.EIGHTBITS,
-    parity=serial.PARITY_NONE,
-    stopbits=serial.STOPBITS_ONE,
-    xonxoff=1,
-    rtscts=0,
-)
-API_SERVERS = (
-    # You can add multiple hosts here... just uncomment the line below.
-    ('http://HOST-OR-IP-1/api/v1/datalogger/dsmrreading', 'APIKEY-1'),
-    # ('http://HOST-OR-IP-2/api/v1/datalogger/dsmrreading', 'APIKEY-2'),
-)
-
-
-def read_serial_port(port, baudrate, bytesize, parity, stopbits, xonxoff, rtscts, **kwargs):  # noqa: C901
+def read_serial_port(port, baudrate, bytesize, parity, stopbits, xonxoff, rtscts, timeout, **kwargs):  # noqa: C901
     """
     Opens the serial port, keeps reading until we have a full telegram and yields the result to preserve the connection.
     """
-    logging.info('[%s] Opening serial port: %s', datetime.datetime.now(), port)
+    logging.info('[%s] Opening serial port -> %s', datetime.datetime.now(), port)
     serial_handle = serial.Serial(
         port=port,
         baudrate=baudrate,
@@ -48,7 +29,7 @@ def read_serial_port(port, baudrate, bytesize, parity, stopbits, xonxoff, rtscts
         stopbits=stopbits,
         xonxoff=xonxoff,
         rtscts=rtscts,
-        timeout=20,  # Max time to wait for data.
+        timeout=timeout,
     )
 
     telegram_start_seen = False
@@ -85,41 +66,143 @@ def read_serial_port(port, baudrate, bytesize, parity, stopbits, xonxoff, rtscts
             yield buffer
 
 
-def send_telegram_to_remote_dsmrreader(telegram, api_url, api_key):
+def read_network_socket(host, port):
+    """ Opens a network socket and keeps reading until we have a full telegram. """
+    MAX_SELECT_TIMEOUT = 0.5
+    MAX_BYTES_RECV = 1024
+    MAX_BUFFER = 10 * MAX_BYTES_RECV
+
+    logging.info('[%s] Opening network socket -> %s:%d', datetime.datetime.now(), host, port)
+    socket_handle = socket.socket(
+        socket.AF_INET,  # Hardcoded ipv4, for now
+        socket.SOCK_STREAM
+    )
+
+    try:
+        socket_handle.connect((host, port))
+    except Exception as error:
+        raise RuntimeError('Failed to connect to network socket: {}', error)
+
+    buffer = ''
+
+    while True:
+        rlist, _, _ = select.select([socket_handle], [], [], MAX_SELECT_TIMEOUT)
+
+        if not rlist:
+            continue
+
+        # Usually a bad sign.
+        if len(buffer) > MAX_BUFFER:
+            logging.error(
+                '[%s] Cleared buffer after reaching max size of %d Bytes!',
+                datetime.datetime.now(),
+                MAX_BUFFER
+            )
+            buffer = ''
+
+        incoming_bytes = socket_handle.recv(MAX_BYTES_RECV)
+        data = str(incoming_bytes, 'latin_1')
+        buffer += data
+
+        # Just add data to the buffer until we detect a telegram in it. Should work for 99% of the cases.
+        match = re.search(r'(\/.+\![A-Z0-9]{4})', buffer, re.DOTALL)
+
+        if not match:
+            continue
+
+        buffer = ''
+        yield match.group(1)
+
+
+def _send_telegram_to_remote_dsmrreader(telegram, api_url, api_key, timeout):
     """ Registers a telegram by simply sending it to the application with a POST request. """
+    logging.debug('[%s] Sending telegram to API: %s', datetime.datetime.now(), api_url)
     response = requests.post(
         api_url,
         headers={'Authorization': 'Token {}'.format(api_key)},
         data={'telegram': telegram},
-        timeout=60,  # Prevents this script from hanging indefinitely when the server or network is unavailable.
+        timeout=timeout,  # Prevents this script from hanging indefinitely when the server or network is unavailable.
     )
 
     if response.status_code != 201:
         logging.error('[%s] API error: HTTP %d - %s', datetime.datetime.now(), response.status_code, response.text)
+        return
+
+    logging.debug('[%s] API response OK: Telegram received successfully', datetime.datetime.now())
 
 
-def main():
-    """ Entrypoint for command line execution. """
-    logging.getLogger().setLevel(logging.INFO)
+def _initialize_logging():
+    logging_level = logging.INFO
+
+    if decouple.config('DATALOGGER_DEBUG_LOGGING', default=False, cast=bool):
+        logging_level = logging.DEBUG
+
+    logging.getLogger().setLevel(logging_level)
+
+
+def main():  # noqa: C901
+    """ Entrypoint for command line execution only. """
+    _initialize_logging()
     logging.info('[%s] Starting...', datetime.datetime.now())
 
-    for telegram in read_serial_port(**SERIAL_SETTINGS):
-        logging.info('[%s] Telegram read', datetime.datetime.now())
+    # Settings.
+    DATALOGGER_SETTINGS = {}
+    DATALOGGER_TIMEOUT = decouple.config('DATALOGGER_TIMEOUT', default=20, cast=float)
+    DATALOGGER_SLEEP = decouple.config('DATALOGGER_SLEEP', default=0.5, cast=float)
+    DATALOGGER_INPUT_METHOD = decouple.config('DATALOGGER_INPUT_METHOD')
+    DATALOGGER_API_HOSTS = decouple.config('DATALOGGER_API_HOSTS', cast=decouple.Csv(post_process=tuple))
+    DATALOGGER_API_KEYS = decouple.config('DATALOGGER_API_KEYS', cast=decouple.Csv(post_process=tuple))
 
-        for current_server in API_SERVERS:
-            current_api_url, current_api_key = current_server
-            logging.info('[%s] Sending telegram to: %s', datetime.datetime.now(), current_api_url)
+    if not DATALOGGER_API_HOSTS or not DATALOGGER_API_KEYS:
+        raise RuntimeError('API_HOSTS or API_KEYS not set')
+
+    if len(DATALOGGER_API_HOSTS) != len(DATALOGGER_API_KEYS):
+        raise RuntimeError('The number of API_HOSTS and API_KEYS given do not match each other')
+
+    if DATALOGGER_INPUT_METHOD == 'serial':
+        DATALOGGER_SETTINGS = dict(
+            port=decouple.config('DATALOGGER_SERIAL_PORT'),
+            baudrate=decouple.config('DATALOGGER_SERIAL_BAUDRATE', cast=int),
+            bytesize=serial.EIGHTBITS,
+            parity=serial.PARITY_NONE,
+            stopbits=serial.STOPBITS_ONE,
+            xonxoff=1,
+            rtscts=0,
+            timeout=DATALOGGER_TIMEOUT
+        )
+    elif DATALOGGER_INPUT_METHOD == 'ipv4':
+        DATALOGGER_SETTINGS = dict(
+            host=decouple.config('DATALOGGER_NETWORK_HOST'),
+            port=decouple.config('DATALOGGER_NETWORK_PORT', cast=int),
+        )
+
+    try:
+        input_function = {
+            'serial': read_serial_port,
+            'ipv4': read_network_socket,
+        }[DATALOGGER_INPUT_METHOD]
+    except KeyError:
+        raise RuntimeError('Unsupported DATALOGGER_INPUT_METHOD')
+
+    for telegram in input_function(**DATALOGGER_SETTINGS):
+        logging.debug('[%s] Telegram read: %s', datetime.datetime.now(), telegram)
+
+        for current_server_index in range(len(DATALOGGER_API_HOSTS)):
+            current_api_host = DATALOGGER_API_HOSTS[current_server_index]
+            current_api_url = '{}/api/v1/datalogger/dsmrreading'.format(current_api_host)
+            current_api_key = DATALOGGER_API_KEYS[current_server_index]
 
             try:
-                send_telegram_to_remote_dsmrreader(
+                _send_telegram_to_remote_dsmrreader(
                     telegram=telegram,
                     api_url=current_api_url,
-                    api_key=current_api_key
+                    api_key=current_api_key,
+                    timeout=DATALOGGER_TIMEOUT,
                 )
             except Exception as error:
                 logging.exception(error)
 
-        time.sleep(SLEEP)
+        time.sleep(DATALOGGER_SLEEP)
 
 
 if __name__ == '__main__':  # pragma: no cover